Extract serialization reflect module
