+++ circuit-serialization-reflect/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=circuit-serialization-reflect
+POM_NAME=Circuit (Serialization Reflect)
+POM_DESCRIPTION=Circuit (Serialization Reflect)
